fixed win32 Build fixed some files operations could display annoying system message box on win32 started to look at drag and drop problems fixed some build warnings
git-svn-id: svn://libnui.net/nui/trunk/nui3@4642 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ src/AudioSamples/nuiSampleReader.cpp
@@ -39,7 +39,7 @@ bool nuiSampleReader::Init()
 
 uint32 nuiSampleReader::GetPosition() const
 {
-  return mPosition;
+  return (uint32)mPosition;
 }
 
 void nuiSampleReader::SetPosition(uint32 position)
